Remove: a trash file
